continueing rearranging the project structure as a package...
