+++ run.py
@@ -1,3 +1,4 @@
-from h5temperature import main
+if __name__ == '__main__':
+	from h5temperature import main
 
-main()
+	main()
